chore(tests): consolidate duplicate test suites (#143)
Remove three duplicate test files and fold their unique assertions into
the co-located suites:

- tests/github.test.ts supersedes lib/github.test.ts; ported the
  concurrency gate test and the full-network-failure fallback test
- lib/games/pong/game.test.ts absorbs the boundary-clamp, particle
  expiry, and exact ball-position assertions from tests/game.test.ts
- tests/browser-utils.test.tsx duplicated cn/getLanguageColor/
  useScrollSpy coverage; the smooth-scroll assertions now live in
  lib/smooth-scroll.test.ts (co-located with the module)
- tests/renderer.test.ts moved to lib/games/pong/renderer.test.ts

Net: -3 test files, coverage unchanged (96.58% statements / 97.00%
branches), 105 tests passing.
