+++ lib/games/pong/game.test.ts
@@ -1,5 +1,6 @@
-import { describe, expect, it } from 'bun:test'
+import { describe, expect, it, spyOn } from 'bun:test'
 import { createGame, updateGame } from './game'
+import { PARTICLE_COUNT, PARTICLE_LIFE } from './constants'
 import type { PongColors } from './types'
 
 const colors: PongColors = {
@@ -21,6 +22,8 @@ describe('createGame', () => {
     // 2 vertical (left/right) + 2 horizontal (top/bottom)
     expect(game.paddles.filter((p) => p.isVertical)).toHaveLength(2)
     expect(game.paddles.filter((p) => !p.isVertical)).toHaveLength(2)
+    // Ball starts at 80% width / 30% height with the scaled diagonal velocity
+    expect(game.ball).toMatchObject({ x: 800, y: 180, dx: -8, dy: 8 })
   })
 
   it('places vertical paddles on the left and right edges', () => {
@@ -87,5 +90,54 @@ describe('updateGame', () => {
     }
     expect(scored).toBe(true)
     expect(game.score).toBeGreaterThan(before)
+  })
+
+  it('clamps the ball to every boundary on exact-edge collisions', () => {
+    const game = createGame(400, 300, colors, ['A', 'B'])
+    game.pixels.forEach((pixel) => {
+      pixel.hit = true
+    })
+
+    game.ball.y = game.ball.radius - 1
+    game.ball.dy = -5
+    updateGame(game)
+    expect(game.ball.y).toBe(game.ball.radius)
+    expect(game.ball.dy).toBeGreaterThan(0)
+
+    game.ball.x = game.width - game.ball.radius + 1
+    game.ball.dx = 5
+    updateGame(game)
+    expect(game.ball.x).toBe(game.width - game.ball.radius)
+    expect(game.ball.dx).toBeLessThan(0)
+  })
+
+  it('marks a pixel hit, updates the score, and expires particles', () => {
+    spyOn(Math, 'random').mockReturnValue(0.5)
+    const game = createGame(600, 400, colors, ['A', 'B'])
+    const target = game.pixels[0]
+    target.x = game.width / 2
+    target.y = game.height / 2
+
+    game.ball.x = target.x + target.size / 2
+    game.ball.y = target.y + target.size / 2
+    game.ball.dx = 0
+    game.ball.dy = 0
+
+    updateGame(game)
+
+    expect(target.hit).toBe(true)
+    expect(game.score).toBe(1)
+    expect(game.particles).toHaveLength(PARTICLE_COUNT)
+
+    game.pixels.forEach((pixel) => {
+      pixel.hit = true
+    })
+    game.ball.x = game.width / 2
+    game.ball.y = 10
+    for (let frame = 1; frame < PARTICLE_LIFE; frame++) {
+      updateGame(game)
+    }
+
+    expect(game.particles).toHaveLength(0)
   })
 })
